``` feat(vending-machine): 实现购物车和结算功能
- 添加购物车功能，包括商品加购、减购和数量显示
- 实现购物车列表显示已加购商品及总价计算
- 增加结算功能模块，包含结算按钮和支付界面
- 修改数据结构，使用双向链表管理购物车商品
- 优化商品卡片布局和交互体验
- 添加二维码支付功能支持
```

+++ figure_vending_machine/goods_mangm.h
@@ -8,7 +8,7 @@
 
 typedef struct Goods_info_t{    //商品信息结构
     char title[40];
-    uint64_t price;
+    uint32_t price;
     char goods_pic_path[100];
     uint32_t num;
 }goods_info_t;
